pyproject.toml
[project]
name = "tower-sling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "sling[arrow]>=1.4.26",
    "tower>=0.3.35",
]
